requirements.txt
flask
flask-cors
bcrypt
python-dotenv
flask-jwt-extended
gunicorn
pymongo 